pyproject.toml
[tool.poetry]
name = "snipOCR"
version = "0.1.0"
description = "Snipping tool which puts OCR text in clipboard"
authors = ["Prakhar Patel <rajprakhar2000@gmail.com>"]
license = "GPL-3.0-or-later"

[tool.poetry.dependencies]
python = "^3.8"
opencv-python-headless = "^4.5.5"
pytesseract = "^0.3.8"
Pillow = "^9.0.0"
PyQt5 = "^5.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
